feat: Add disconnect on all pages
